be/requirements.txt
elasticsearch>=8.0.0,<9.0.0
urllib3>=1.26.0
tqdm>=4.62.0
fastapi
uvicorn[standard]
pydantic
python-multipart
typing-extensions
python-dotenv
requests
numpy
python-dotenv
json_repair
redis>=4.0.0
aiohttp